backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching & Queues
redis==5.0.1
celery==5.3.4

# AWS Integration
boto3==1.34.0
botocore==1.34.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & Networking
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.2.0
numpy==1.26.0

# Report Generation
reportlab==4.0.8
openpyxl==3.1.2
jinja2==3.1.2
weasyprint==61.2

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.39.1

# Task Scheduling
apscheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.1

# Development
black==23.11.0
isort==5.13.2
mypy==1.7.1